ZO-358: Group bootstrap-related code together
We now might want to move zeit.cms.application -> zeit.cms.browser.application
